webapp/requirements.txt
django==3.1.14
whitenoise==6.0.0
psycopg2==2.9.3
pillow==9.0.1
numpy==1.22.3
requests==2.27.1
